@@ -1,7 +1,7 @@
 export  const theme = {
     colors: {
         black: '#000000',
-        white: '#FFFFFF',
+        white: 'rgba(255, 255, 255, 1)',
         emperor: '#555555',
         silver: '#CCCCCC',
         teal: '#008080',
